factor cache & chunk def functions to separate translation units
