+++ src/cpp/session/CMakeLists.txt
@@ -155,6 +155,8 @@ set (SESSION_SOURCE_FILES
    modules/rmarkdown/SessionRMarkdown.cpp
    modules/rmarkdown/SessionRmdNotebook.cpp
    modules/rmarkdown/SessionRnbParser.cpp
+   modules/rmarkdown/NotebookCache.cpp
+   modules/rmarkdown/NotebookChunkDefs.cpp
    modules/rmarkdown/NotebookPlots.cpp
    modules/rmarkdown/RMarkdownPresentation.cpp
    modules/shiny/SessionShiny.cpp
