+++ src/cpp/session/modules/rmarkdown/NotebookCache.hpp
@@ -1,5 +1,5 @@
 /*
- * SessionRmdNotebook.hpp
+ * NotebookCache.hpp
  *
  * Copyright (C) 2009-16 by RStudio, Inc.
  *
@@ -14,44 +14,31 @@
  */
 
 
-#ifndef SESSION_RMARKDOWN_NOTEBOOK_HPP
-#define SESSION_RMARKDOWN_NOTEBOOK_HPP
-
-#include <ctime>
-#include <boost/signals.hpp>
-#include <core/json/Json.hpp>
-
-#define kChunkLibDir "lib"
+#ifndef SESSION_NOTEBOOK_CACHE_HPP
+#define SESSION_NOTEBOOK_CACHE_HPP
 
 namespace rstudio {
 namespace core {
-   class Error;
    class FilePath;
+   class Error;
 }
 }
- 
+
 namespace rstudio {
 namespace session {
 namespace modules {
 namespace rmarkdown {
 namespace notebook {
 
-core::Error initialize();
+core::FilePath chunkCacheFolder(const std::string& docPath, 
+      const std::string& docId, const std::string& contextId);
 
-struct Events : boost::noncopyable
-{
-   // Document {0}, chunk {1} from context id {3} execution completed
-   boost::signal<void(const std::string&, const std::string&,
-                      const std::string&)> 
-                onChunkExecCompleted;
+core::FilePath chunkCacheFolder(const std::string& docPath, 
+      const std::string& docId);
 
-   // Document {0}, chunk {1} had console output of type {2} and text {3}
-   boost::signal<void(const std::string&, const std::string&, int, 
-                const std::string&)>
-                onChunkConsoleOutput;
-};
+core::Error ensureCacheFolder(const core::FilePath& folder);
 
-Events& events();
+core::Error initCache();
 
 } // namespace notebook
 } // namespace rmarkdown
@@ -59,4 +46,4 @@ Events& events();
 } // namespace session
 } // namespace rstudio
 
-#endif // SESSION_RMARKDOWN_NOTEBOOK_HPP
+#endif
